requirements.txt
numpy==1.26.0
torch==2.1.0
torchvision==0.16.0
pillow==10.0.1
pandas==2.1.1
streamlit==1.28.1